Various uses of atoi should have been atoll (for int64_t)

+++ src/wallet/wallet.h
@@ -132,7 +132,7 @@ static inline void ReadOrderPos(int64_t &nOrderPos, mapValue_t &mapValue) {
         return;
     }
 
-    nOrderPos = std::atoi(mapValue["n"].c_str());
+    nOrderPos = std::atoll(mapValue["n"].c_str());
 }
 
 static inline void WriteOrderPos(const int64_t &nOrderPos,
